Criei o componente DataPanel e importei ele no Checkout e profile
