+++ src/globalStyle/style.js
@@ -1,5 +1,5 @@
 // src/GlobalStyle.js
-
+import styled from 'styled-components';
 import { createGlobalStyle } from "styled-components";
 
 const GlobalStyle = createGlobalStyle`
@@ -28,3 +28,52 @@ const GlobalStyle = createGlobalStyle`
 `;
 
 export default GlobalStyle;
+
+
+export const GlobalButtonLayout = styled.button`
+  width: 200px;
+  height: 50px;
+  padding: 12px 20px;
+  border-radius: 6px;
+  border: none;
+  cursor: pointer;
+  font-size: 15px;
+  transition: background-color 0.3s ease;
+
+  &.greenBtn {
+    background-color: #4caf50;
+    color: white;
+
+    &:hover {
+      background-color: #45a049;
+    }
+  }
+
+  &.redBtn {
+    background-color: #a83232;
+    color: white;
+
+    &:hover {
+      background-color: #9f2929;
+    }
+  }
+
+  &.yellowBtn {
+    background-color: #ffbe55;
+    color: #fff;
+
+    &:hover {
+      background-color: #ffab33;
+    }
+  }
+
+  &.whiteBtn {
+    background-color: #e0e0e0;
+    color: #000;
+
+    &:hover {
+      background-color: #f0f0f0;
+    }
+  }
+`;
+
