+++ src/pages/Checkout/index.jsx
@@ -1,34 +1,30 @@
 import React from 'react';
 import { Link } from "react-router-dom";
-import logoHeader from "../../assets/logoheader.jpeg";
-import GlobalStyle from "../../globalStyle/style.js"; // Importando o estilo global
-import styles, { CheckoutContainer, Header, FormSection, InputGroup, AddressSection, FreightAndSummary, FreightSection, SummarySection, ButtonGroup } from './styles';
+import GlobalStyle, { GlobalButtonLayout } from "../../globalStyle/style.js"; // Importando o estilo global
+import styles, { CheckoutContainer, FreightAndSummary, FreightSection, SummarySection } from './styles';
+import Navbar from "../../components/navBar.jsx";
+import DataPanel from "../../components/dataPanel.jsx"
 
 const Checkout = () => {
   return (
-    <div className='body' style={styles.body}>
-      <GlobalStyle />
-      <HeaderPage />
-      <Panel />
+    <div>
+      <Navbar />
+      <div className='body' style={styles.body}>
+        <GlobalStyle />        
+        <FullCheckout />
+      </div>
     </div>
   );
 };
 
 export default Checkout;
 
-function Panel() {
+
+
+function FullCheckout() {
   return (
     <CheckoutContainer>
-      <h2>Meus Dados</h2>
-      <div className='meus-dados'>
-        <div className="form-container">
-          <DataForm />
-          <AddressForm />
-        </div>
-        <ButtonContainer />
-      </div>
-
-
+      <DataPanel />
       <FreightAndSummary>
         <FreightSection>
           <h3>Frete</h3>
@@ -57,97 +53,10 @@ function Panel() {
           <p>Total: <span>R$ 128,19</span></p>
         </SummarySection>
       </FreightAndSummary>
-
-      <ButtonGroup>
-        <Link to="/">
-          <button className="back">Voltar</button>
-        </Link>
-        <Link to="/">
-          <button className="continue">Continuar</button>
-        </Link>
-      </ButtonGroup>
+      <div className='SumaryBtn'>
+        <GlobalButtonLayout className="whiteBtn">Voltar</GlobalButtonLayout>
+        <GlobalButtonLayout className="yellowBtn">Continuar</GlobalButtonLayout>
+      </div>
     </CheckoutContainer>
-  );
-}
-
-//temporário <<<<<<<<<<<<<<<<<<<<<<<>>>>>>>>>>>>>>>>>>>>>
-function HeaderPage() {
-  return (
-    <Header>
-      <img src={logoHeader} alt="Logo Cabeçalho" />
-    </Header>
-  );
-}
-
-function DataForm() {
-  return (
-    <FormSection>
-      <h3>Dados Pessoais:</h3>
-      <InputGroup>
-        <label>Nome</label>
-        <input type="text" placeholder="Luíza Fernanda Ribeiro dos Santos" />
-      </InputGroup>
-      <InputGroup>
-        <label>Telefone Celular</label>
-        <input type="text" placeholder="(95) 2547-1891" />
-      </InputGroup>
-      <InputGroup>
-        <label>Whatsapp</label>
-        <input type="text" placeholder="(95) 99912-6177" />
-      </InputGroup>
-      <InputGroup>
-        <label>E-mail</label>
-        <input type="email" placeholder="luiza.ferbi@mail.com" />
-      </InputGroup>
-      <InputGroup>
-        <label>Data de Nascimento</label>
-        <input type="date" />
-      </InputGroup>
-    </FormSection>
-  );
-}
-
-function AddressForm() {
-  return (
-    <AddressSection>
-      <h3>Endereço de Entrega:</h3>
-      <InputGroup>
-        <label>CEP</label>
-        <input type="text" placeholder="79900-502" />
-      </InputGroup>
-      <InputGroup>
-        <label>Logradouro</label>
-        <input type="text" placeholder="Rua Madalena Manvailer" />
-      </InputGroup>
-      <InputGroup>
-        <label>Número</label>
-        <input type="text" placeholder="396" />
-      </InputGroup>
-      <InputGroup>
-        <label>Complemento</label>
-        <input type="text" placeholder="Apto 4" />
-      </InputGroup>
-      <InputGroup>
-        <label>Bairro</label>
-        <input type="text" placeholder="Parque dos Ipês III" />
-      </InputGroup>
-      <InputGroup>
-        <label>Cidade</label>
-        <input type="text" placeholder="Ponta Porã" />
-      </InputGroup>
-      <InputGroup>
-        <label>UF</label>
-        <input type="text" placeholder="MS" />
-      </InputGroup>
-    </AddressSection>
-  );
-}
-
-function ButtonContainer() {
-  return (
-    <ButtonGroup>
-      <button className="save">Salvar</button>
-      <button className="logout">Logout</button>
-    </ButtonGroup>
   );
 }
